services/search-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
exa-py
google-genai
pymongo
python-dotenv
pytest
pytest-asyncio
httpx
python-jose[cryptography] 
